test: refactor the code that waits for service state etc per review
